ensure that chromium import is restricted to its own file so individual function packaging works
