+++ api/.swcrc
@@ -66,6 +66,9 @@
             "link/*": [
                 "src/components/link/*"
             ],
+            "pdf/*": [
+                "src/components/pdf/*"
+            ],
             "publication/*": [
                 "src/components/publication/*"
             ],
