+++ api/serverless-config-default.yml
@@ -45,7 +45,7 @@ functions:
                 method: GET
                 cors: true
     getPDF:
-        handler: dist/src/components/publication/routes.getPDF
+        handler: dist/src/components/pdf/routes.getPublicationPDF
         memorySize: 2048 # https://github.com/alixaxel/chrome-aws-lambda#:~:text=You%20should%20allocate%20at%20least%20512%20MB%20of%20RAM%20to%20your%20Lambda%2C%20however%201600%20MB%20(or%20more)%20is%20recommended
         events:
             - http:
@@ -530,7 +530,6 @@ package:
         - '!src/**/__tests__/**'
         - '!node_modules/.prisma/client/libquery_engine-debian-openssl-3.0.x.so.node'
         - '!node_modules/.prisma/client/libquery_engine-rhel-openssl-1.0.x.so.node'
-        - '!node_modules/.prisma/client/libquery_engine-rhel-openssl-3.0.x.so.node'
         - '!node_modules/@prisma/client/runtime/*.wasm'
         - '!node_modules/@prisma/engines/**'
         - '!node_modules/@sparticuz/chromium/**'
@@ -563,7 +562,6 @@ package:
         - '!src/**/__tests__/**'
         - '!node_modules/.prisma/client/libquery_engine-debian-openssl-3.0.x.so.node'
         - '!node_modules/.prisma/client/libquery_engine-rhel-openssl-1.0.x.so.node'
-        - '!node_modules/.prisma/client/libquery_engine-rhel-openssl-3.0.x.so.node'
         - '!node_modules/@prisma/client/runtime/*.wasm'
         - '!node_modules/@prisma/engines/**'
         - '!node_modules/@sparticuz/chromium/**'
