+++ api/src/components/pdf/schema/getPublicationPDF.ts
