+++ api/src/components/pdf/schema/index.ts
@@ -1,0 +1,1 @@
+export { default as getPublicationPDF } from './getPublicationPDF';
