+++ api/src/components/publication/routes.ts
@@ -24,8 +24,3 @@ export const getPublicationLinks = middy(publicationController.getLinksForPublic
     .use(middleware.doNotWaitForEmptyEventLoop({ runOnError: true, runOnBefore: true, runOnAfter: true }))
     .use(middleware.httpJsonBodyParser())
     .use(middleware.authentication(true));
-
-export const getPDF = middy(publicationController.getPDF)
-    .use(middleware.doNotWaitForEmptyEventLoop({ runOnError: true, runOnBefore: true, runOnAfter: true }))
-    .use(middleware.httpJsonBodyParser())
-    .use(middleware.validator(publicationSchema.getPDF, 'pathParameters'));
