+++ api/src/components/publication/schema/index.ts
@@ -1,3 +1,2 @@
 export { default as createBody } from './createBody';
 export { default as createQueryStringParameters } from './createQueryStringParameters';
-export { default as getPDF } from './getPDF';
